Refactor: move timeline chart to own library
