+++ bnovate/bnovate/report/work_order_planning/work_order_planning.js
@@ -4,6 +4,7 @@
 
 frappe.require("/assets/bnovate/js/lib/gcharts/loader.js")
 frappe.require("/assets/bnovate/js/modals.js")  // provides bnovate.modals
+frappe.require("/assets/bnovate/js/charts.js")  // provides bnovate.modals
 
 frappe.query_reports["Work Order Planning"] = {
     filters: [
@@ -30,7 +31,7 @@ frappe.query_reports["Work Order Planning"] = {
         bnovate.modals.attach_report_modal("reportModal");
     },
     after_datatable_render(datatable) {
-        draw_google_chart(this.report);
+        bnovate.charts.draw_timeline_chart(this.report, build_wo_dt);
 
         // Activate tooltips on columns
         $(function () {
@@ -105,84 +106,7 @@ function projected_stock_link(item_code, warehouse, item_name) {
         });
 }
 
-// Popover is the popup shown when clicking a bar on the timeline chart.
-// #sacrificial is a little 0x0 px square created at the location clicked.
-function dismiss_popover() {
-    $('#sacrificial').popover('hide');
-    $('#sacrificial').remove();
-}
-
-function draw_google_chart(report) {
-    report.$chart.html(`
-            <div class="chart-container">
-                <div id="timeline" class="report-chart">Timeline</div>
-            </div>
-    `);
-
-    // Track mouse position in chart to draw popover at correct location
-    report.mousePos = { x: 0, y: 0 };
-    let container = document.getElementById("timeline");
-    container.addEventListener("mousemove", (e) => {
-        report.mousePos.x = e.offsetX;
-        report.mousePos.y = e.offsetY;
-    });
-
-    google.charts.load('current', { 'packages': ['timeline'] });
-    google.charts.setOnLoadCallback(drawChart);
-    function drawChart() {
-        let chart = new google.visualization.Timeline(container);
-        report.chart = chart;
-
-        report.chart_dt = build_google_dt(report);
-
-        // Show a popover when a timeline element is clicked. Popovers in bootstrap 3 don't
-        // delete nicely, so I made a workaround where I create a sacrificial div centered on the
-        // click and bind the popover to this div.
-        google.visualization.events.addListener(chart, 'select', () => {
-
-            dismiss_popover();
-
-            $(container).append(
-                `<div id="sacrificial" style="left: ${report.mousePos.x}px; top: ${report.mousePos.y}px"></div>`
-            )
-
-            let row = chart.getSelection()[0].row;
-            let contents = report.chart_dt.getValue(row, 2);
-
-            $("#sacrificial").popover({
-                container: "body",
-                title: "My fancy popover",
-                html: true,
-                content: contents,
-                placement: "bottom",
-            }).popover('show');
-
-        });
-
-        // Any click outside the popover dismisses it. May interfere with other popovers...
-        document.addEventListener("click", (event) => {
-            const popover = $(".popover");
-            if (popover.is(event.target) || popover.has(event.target).length) {
-                // Ignore if click target is inside a popover
-                return;
-            }
-            dismiss_popover();
-        });
-        frappe.route.on('change', () => { dismiss_popover() });
-
-        chart.draw(report.chart_dt, {
-            height: 250,
-            fontName: "Open Sans",
-            tooltip: {
-                // isHtml: true,
-                trigger: 'both', // disable it
-
-            },
-        });
-    }
-}
-
-function build_google_dt(report) {
+function build_wo_dt(report) {
     var dataTable = new google.visualization.DataTable();
 
     dataTable.addColumn({ type: 'string', id: 'Workstation' });
